backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.4.2
httpx==0.25.0
aiohttp==3.9.0
websockets==12.0
redis==5.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
PyJWT==2.8.1
email-validator==2.1.0
alembic==1.13.0
psycopg2-binary==2.9.9
requests==2.31.0
